+++ .travis.yml
@@ -129,7 +129,7 @@ script:
                 -DCUTTER_ENABLE_JUPYTER=ON
                 -DCUTTER_ENABLE_QTWEBENGINE=OFF
                 -DCUTTER_ENABLE_CRASH_REPORTS=true
-                -DBREAKPAD_FRAMEWORK_DIR=$BREAKPAD_FRAMEWORK_DIR
+                -DBREAKPAD_FRAMEWORK_DIR="$BREAKPAD_FRAMEWORK_DIR"
                 ../src &&
             make -j4;
        fi
