+++ src/common/BugReporting.h
@@ -1,0 +1,6 @@
+#ifndef CRASHREPORTING_H
+#define CRASHREPORTING_H
+
+void openIssue();
+
+#endif // CRASHREPORTING_H
