+++ src/common/CrashHandler.h
@@ -9,6 +9,5 @@
 */
 void initCrashHandler();
 
-void openIssue();
 
 #endif // CRASH_HANDLER_H
