+++ src/core/CutterDescriptions.h
@@ -9,6 +9,7 @@
 #include <QList>
 #include <QStringList>
 #include <QMetaType>
+#include "core/CutterCommon.h"
 
 struct FunctionDescription {
     RVA offset;
